More cleanup of the wscat copied code
