+++ tichu-clients/tichu-cli-client/src/ws-client.ts
@@ -3,13 +3,14 @@ import inquirer from "inquirer";
 import { Cards } from "./cards";
 import { GameMessage, JoinParam, Message } from "./actions";
 import { GameOpts } from "./startup";
+import { Console } from "./console";
 
 export class WSTichuClient {
     private console: Console;
     private _webSocket: WebSocket | undefined;
 
     constructor(readonly opts: GameOpts) {
-        this.console = new Console(this.send, this.close);
+        this.console = new Console();
     }
 
     connect(url: string) {
@@ -29,15 +30,12 @@ export class WSTichuClient {
         return "Done";
     }
 
-
     promptForCards = () => {
         inquirer
             .prompt([
                 {
                     message:
-                        "Pick cards, hit enter to play them. Just hit enter to pass.",
-                    prefix: "<this is the prefix>",
-                    suffix: "<this is the suffix>",
+                        "Pick cards with space, hit enter to play them. Just hit enter to pass.",
                     type: "checkbox",
                     name: "cards",
                     choices: Cards.sort(() => 0.5 - Math.random()).slice(0, 14),
@@ -64,7 +62,6 @@ export class WSTichuClient {
 
     send = (msg: Message) => {
         this.ws().send(JSON.stringify(msg));
-        this.console.prompt();
     };
 
     private receive(data: WebSocket.Data) {
@@ -77,7 +74,7 @@ export class WSTichuClient {
     }
 
     close = () => {
-        this.ws().close(); //closeStatusCode, closeReason);
+        this.ws().close();
     };
 
     private ws(): WebSocket {
@@ -99,37 +96,6 @@ export class WSTichuClient {
                 Console.Colors.Green
             );
             this.onConnect();
-            // setup line-handlers on console object:
-            /*
-            wsConsole.on("line", data => {
-                switch (data) {
-                    case "/ping":
-                        ws.ping(noop);
-                        break;
-                    case "/pong":
-                        ws.pong(noop);
-                        break;
-                    case "/close": {
-                        let closeStatusCode = 1000;
-                        let closeReason = "";
-                        if (toks.length >= 2) {
-                            closeStatusCode = parseInt(toks[1]);
-                        }
-                        if (toks.length >= 3) {
-                            closeReason = toks.slice(2).join(" ");
-                        }
-                        if (closeReason.length > 0) {
-                            ws.close(closeStatusCode, closeReason);
-                        } else {
-                            ws.close(closeStatusCode);
-                        }
-                        break;
-                    default:
-                        ws.send(data);
-                    }
-                        wsConsole.prompt();
-                }
-            );*/
         });
 
         ws.on("close", (code, reason) => {
@@ -173,85 +139,6 @@ export class WSTichuClient {
 
         return ws;
     }
-}
-
-// copied from wscat
-class Console {
-    // extends EventEmitter {
-    // private stdin: NodeJS.ReadStream;
-    private stdout: NodeJS.WriteStream;
-    // private readlineInterface: Interface;
-    private _resetInput: () => void;
-
-    constructor(onLine: (line: string) => void, onClose: () => void) {
-        // super();
-
-        // this.stdin = process.stdin;
-        this.stdout = process.stdout;
-
-        // this.readlineInterface = readline.createInterface(
-        //     this.stdin,
-        //     this.stdout
-        // );
-
-        // this.readlineInterface.on("line", onLine).on("close", onClose);
-
-        this._resetInput = () => {
-            this.clear();
-        };
-    }
-
-    static get Colors() {
-        return {
-            Red: "\u001b[31m",
-            Green: "\u001b[32m",
-            Yellow: "\u001b[33m",
-            Blue: "\u001b[34m",
-            Default: "\u001b[39m"
-        };
-    }
-
-    static get Types() {
-        return {
-            Incoming: "< ",
-            Control: "",
-            Error: "error: "
-        };
-    }
-
-    prompt() {
-        // this.readlineInterface.prompt();
-    }
-
-    print(type: string, msg: string, color: string) {
-        // if (tty.isatty(1)) {
-        //     this.clear();
-        //
-        //     if (program.execute) color = type = "";
-        //     else if (!program.color) color = "";
-
-        this.stdout.write(color + type + msg + Console.Colors.Default + "\n");
-        this.prompt();
-        // } else if (type === Console.Types.Incoming) {
-        //     this.stdout.write(msg + "\n");
-        // } else {
-        //     // is a control message and we're not in a tty... drop it.
-        // }
-    }
-
-    clear() {
-        // if (tty.isatty(1)) {
-        this.stdout.write("\u001b[2K\u001b[3D");
-        // }
-    }
-
-    // pause() {
-    //     this.stdin.on("keypress", this._resetInput);
-    // }
-    //
-    // resume() {
-    //     this.stdin.removeListener("keypress", this._resetInput);
-    // }
 }
 
 export type Status = "Done" | "Not done";
